Add HA calibration service
Adding support for calibration service of power, current and voltage through Home Assistant to knightsbridge-cu9kw.yaml

To perform the calibration, put a known load on the plug and in HA got to Developer Tools -> Actions -> <device_name>_calibrate_current and enter the correct amount for the known load. Repeat for voltage and power.